ARX(1) model
The fit of the ARX model deviates from what we would expect. The trend parameter is much lower than without AR(1) term. In R we find more plausible results. The PyFlux package in Python defines a full probabilistic model (Bayesian approach). We have to dig deeper to see how to incorporate independent variables to the fit in StatsModels.